refactor(app): extract coordinators from useRuntimeConsole into per-domain hooks
The action functions in `useRuntimeConsole.ts` move into seven per-domain
coordinator hooks under `ui/src/app/state/coordinators/`. After this PR,
useRuntimeConsole is a thin composer: slice hooks, coordinator hooks,
derived state + memoized selectors, cross-slice effects, and the same
`{state, actions}` viewmodel it returns today.

The hook drops from 1642 to 529 LOC; the new coordinators are
chat (936 LOC, the biggest bundle: submission, lifecycle, files,
approvals, targeting), providers (203 LOC, CRUD + capability overrides),
dashboard (142 LOC, loadDashboard fan-out + refresh helpers),
settings (63 LOC, runSettingsMutation + notice helpers),
agentAdapters (54 LOC, probe + credential), policy (41 LOC, rule CRUD),
and retention (26 LOC, notice-banner-wired run).

The viewmodel returned by `useRuntimeConsole()` is byte-identical:
ui/src/app/useRuntimeConsole.test.tsx is unchanged, 939 / 939 pass.

Cross-coordinator structure: each hook accepts a small parameter bag of
non-slice dependencies (settings setters, sibling-coordinator handles).
The dashboard → chat → settings cycle (loadDashboard is called by
runSettingsMutation, applyChatSession is called by loadDashboard,
setNoticeMessage is called by chat) is broken with one ref in the
composer: useSettingsActions resolves loadDashboard through a lazy
thunk populated after useDashboardActions runs. No refs cross hooks.
